🔧 update: extract data table components for reuse to $components/custom
